+++ examples/can-send.rs
@@ -1,0 +1,100 @@
+//! Simple CAN example.
+//! Requires a transceiver connected to PB8, 9 (CAN1) or PB5 PB6 (CAN2).
+
+#![no_main]
+#![no_std]
+
+use panic_halt as _;
+
+use bxcan::filter::Mask32;
+use bxcan::{Frame, StandardId};
+use cortex_m_rt::entry;
+use nb::block;
+use stm32f4xx_hal::{can::Can, pac, prelude::*};
+
+#[entry]
+fn main() -> ! {
+    let dp = pac::Peripherals::take().unwrap();
+
+    let rcc = dp.RCC.constrain();
+
+    // To meet CAN clock accuracy requirements an external crystal or ceramic
+    // resonator must be used. The blue pill has a 8MHz external crystal.
+    // Other boards might have a crystal with another frequency or none at all.
+    rcc.cfgr.use_hse(8.mhz()).freeze();
+
+    let gpiob = dp.GPIOB.split();
+    let mut can1 = {
+        let rx = gpiob.pb8.into_alternate_af9();
+        let tx = gpiob.pb9.into_alternate_af9();
+
+        let can = Can::new(dp.CAN1, (tx, rx));
+
+        bxcan::Can::new(can)
+    };
+    can1.configure(|config| {
+        // APB1 (PCLK1): 8MHz, Bit rate: 500kBit/s, Sample Point 87.5%
+        // Value was calculated with http://www.bittiming.can-wiki.info/
+        config.set_bit_timing(0x001c_0000);
+    });
+
+    // Configure filters so that can frames can be received.
+    let mut filters = can1.modify_filters();
+    filters.enable_bank(0, Mask32::accept_all());
+
+    let _can2 = {
+        let tx = gpiob.pb13.into_alternate_af9();
+        let rx = gpiob.pb12.into_alternate_af9();
+
+        let can = Can::new(dp.CAN2, (tx, rx));
+
+        let mut can2 = bxcan::Can::new(can);
+        can2.configure(|config| {
+            // APB1 (PCLK1): 8MHz, Bit rate: 500kBit/s, Sample Point 87.5%
+            // Value was calculated with http://www.bittiming.can-wiki.info/
+            config.set_bit_timing(0x001c_0000);
+        });
+
+        // A total of 28 filters are shared between the two CAN instances.
+        // Split them equally between CAN1 and CAN2.
+        filters.set_split(14);
+        let mut slave_filters = filters.slave_filters();
+        slave_filters.enable_bank(14, Mask32::accept_all());
+        can2
+    };
+
+    // Drop filters to leave filter configuraiton mode.
+    drop(filters);
+
+    // Select the interface.
+    let mut can = can1;
+    //let mut can = can2;
+
+    // Split the peripheral into transmitter and receiver parts.
+    block!(can.enable()).unwrap();
+
+    // Echo back received packages in sequence.
+    // See the `can-rtfm` example for an echo implementation that adheres to
+    // correct frame ordering based on the transfer id.
+    let mut test: [u8; 8] = [0; 8];
+    let mut count: u8 = 0;
+    let id: u16 = 0x500;
+
+    test[1] = 1;
+    test[2] = 2;
+    test[3] = 3;
+    test[4] = 4;
+    test[5] = 5;
+    test[6] = 6;
+    test[7] = 7;
+    loop {
+        test[0] = count;
+        let test_frame = Frame::new_data(StandardId::new(id).unwrap(), test);
+        block!(can.transmit(&test_frame)).unwrap();
+        if count < 255 {
+            count += 1;
+        } else {
+            count = 0;
+        }
+    }
+}